+++ adbnik/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "2.7.0"
+__version__ = "4.0.0"
 APP_TITLE = "Adbnik"